requirements.txt
pickle
numpy
pillow
opencv-python
opencv-contrib-python
